+++ CHANGELOG.md
@@ -4,6 +4,18 @@ We are currently working on porting this changelog to the specifications in
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## Version 1.3.0 -
+
+### Added
+
+* Add `editfile` to the xdev CLI
+
+### Changed
+
+* Support more default editors in `editfile`.
+
+
 ## Version 1.2.0 - Released 2023-04-04 
 
 ### Added